Address JAX score_mod review feedback
